requirements-dev.txt
pytest===3.3.1
django===1.11
django-filter===1.1.0
djangorestframework===3.5.3
pytest-django===3.1.2
factory-boy===2.9.0